Mesh rendering & shaders
---

This commit introduces a stable version of the program where meshes can
be drawn on screen. For now they only show the mesh colored according to
the normal at that point, but it's a start.

Also, the camera does not want to update according when I want it to. So
once a mesh is loaded it may be necessary to click on the 3D viewer, and
press Ctrl+A in order to recenter and widen the camera's radius.