ARAP Integration successful.
---

Now, there needs to be some interaction primitives in order to allow the
user to deform a mesh. For now, there's only a dummy ARAP deformation
performed when pressing Shift+A

But I'd like to investigate why some VAOs are not used in the Scene ...
Might have to do with Thomas' recent code removal from the Scene class.