Visualisation of mesh to coordinates & pairing
---

It is now possible to get the position of the rasterized fragments of
the mesh by clicking on them in the 3D viewer.

Also, when loading a mesh while an image is loaded the user is given the
choice to pair the mesh with a grid or not.

Also, fixed the problem where some spheres would not show at their
designated position in the point/sphere shader program.