Scale the mesh according to the image
---

When a mesh is loaded in while an image was previously loaded, a dialog
asks the user whether or not to 'pair' one of the loaded images and the
mesh. If one is selected, the mesh will be scaled and translated so it
is positioned next to the image, in order to select points on the mesh
and in the image in order to perform a first-guess of the registration
and deformation of the mesh.