GLSL fixes for Intel shader compiler
---

Since it was more strict than Nvidia's GLSL compiler, it spat out a
bunch of errors related to inputs/outputs of the geometry and fragment
shader stages. Those are now fixed !

Also, the extension GL_ARB_explicit_attrib_location is force-enabled for
now, it allows to use 'layout(location)' qualifiers in shaders.