ARAP onto image testing, and failing
---

First few tests of the Mesh/Image deformation using ARAP did not go well
at all. Current suspicion is the transform matrix applied to the mesh.
Should decompose the application of the matrix with the ARAP
computation, in order to first check if the 'alignment' of the mesh and
the image is good enough.

Maybe use the MatrixTransform class ? Could be nice.