Constrained Image->Mesh ARAP done
---

This commit introduces a 'stable-enough' ARAP image to mesh registration. By selecting a series of points on the mesh and selecting the corresponding points on the patient data, the user can perform a rough rigid registration of the anatomical markers of the mesh and image.
This is a proof-of-concept, and the code is _VERY_ WIP. Should not make its way onto anything user-operated.

This commit also removes the default resize of the window to a 1440p framebuffer, and bumps the minimum size to a 720p framebuffer.