Fix fragment depth on ray-tracing
---

For what seems like forever, the fragment depth on the volumetric rendering of a grid would be equal to the fragment depth of the tetrahedron the ray-analysis originated from.

This commit modifies the fragment shader of the volumetric renderer to ensure fragment depth is equal to the depth of the 'virtual voxel' that should be rendered, not the depth of its surrounding acceleration structure. It fixes a _lot_ of sorting issues in the pipeline.