Cage coordinates implemented for ARAP
---

It is now possible to have a curve be deformed by a Mesh which is itself deformed in some way (user, ARAP ...). The computed coordinates are MVC (mean-value-coordinates) but Green coordinates and URAGO (?) coordinates are also implemented.
For now, the curve is overlayed over everything (see new_shaders/base_curve.frag, line 12 where gl_FragDepth is set statically). The rework of the scene graph cannot come soon enough.

The curve to load is an OBJ file, and it should be user-provided.
A default curve and mesh were used for the demo, should see for more cases if it works.