Third draft : draws, but no manipulators
---

Seems like the manipulator origin is not properly computed.

See meshes/operations/arap/mesh_manip_interface.hpp, in the
loadAndInitialize function. It _should_ set modified vertices, but it
doesn't seem like the modifiedvertices vector is filled in properly.
Since whenever the mean of all modified vertices is computed, it is then
divided by 0 which produces a NaN vector.

See :
  - RectangleSelection::mouseReleaseEvent()
    -> calls (through signal) Viewer::rectangleSelection_add
    -> calls MMInterface::select() which enables vertices

After that, _some_ vertices should be set as 'selected_vertices' which
are used in MMInterface::computeManipulatorForSelection() which sets the
origin and size of manipulators.