Working real-time ARAP on meshes
---

Not only is the ARAP deformation framework working in as close to real
time as possible on the machine, I also solved a long-standing bug where
geometry drawn using legacy OpenGL commands would render in black.

It was a combination of a wrongfully enabled GL_COLOR_MATERIAL, a series
of unsupported features enabled using glEnable(), and some more legacy
options enabled by different parts of the program.

The ARAP deformation is working as intended, and should only be reviewed
a bit before merging to the develop/ branch.