First draft of the ARAP deformation program
---

Continuing the work done in the last commit to introduce the ARAP
controller, this commit introduces a new target : arap_deformation

This target sources the arap_main.cpp file as its main(), and creates an
ARAPMainWidget-based program. This allows to customize the program's UI
and behaviour for the purposes of ARAP deformation within my thesis
project.

Also, the GLMDataModel class wwas introduced, and is a way to represent
a vector of glm::vec3 positions using Qt's MVC paradigm. This is still a
work in progress, and it is not included in any compilable targets for
now.