Added checkbox to the ARAPController
---

The deformation is not updated yet, but will be soon.

Also, check if we properly update the position of the mesh vertices in
the ARAPController::applyTransformation_Mesh() function. When selecting
a set of points in the mesh, computing the manipulator moved the points
to the back of the mesh itself.