Change the drawing of the MMInterface
---

With current OpenGL drivers, the GL lists were horribly underperforming
in rendering tasks. So much so that drawing anything with them caused a
good 10x slowdown in rendering speed.

The MMInterface has been changed in order to precompute a vector of
positions for the fixed and moving vertices, and this vector will be
used in the Scene function which draws a set of spheres wuickly using a
'modern' shader pipeline.

This results in nearly no loss in performance using the latest Nvidia
drivers.