More draws in Scene & features in ARAPController
---

Added the Scene::drawColoredPointSpheres_quick which allows to draw a
whole bunch of spheres quite fast, using a shader pipeline and a set
color for all the spheres. Used to draw the anatomical markers on the
mesh and image, and also for the mesh constraints from the MMInterface.

Added more features to the ARAPController, ported over from the POC
earlier in development.