Add explicit setting of voxel dimensions
---

It was not possible before, and it makes sense in this context.

It should be noted that for now, the grids where setting a voxel
dimension explicitely is prohibited throw a runtime error. It might be
better to have the function either :
- return a boolean to indicate if the operation was allowed or not
- return a reference to *this and silently fail when not allowed
- return nothing, throw no error -> fail silently again

Time will tell which is a better alternative.