[WIP] Export image using CImg & clicks
---

CImg does not seem to compile on my laptop due to multiple definitions
of _fseeki64() and _ftelli64() by MinGW 8.0.0 for Qt. Will investigate
on another machine.
So, hid the code behind a small ifdef CIMG_WORKING_ON_MINGW.

Also, re-introduced the ability to click on PlanarViewers to select a
coordinate in 3D space. Coordinates are flipped, but shouldn't be too
hard to reverse.